archive command
renames file to .file

All behaviour is the same except now the file is
hidden from the list command.

This means the file still lives in the same directory
can be decrypted/edited/removed as per normal (plus
being in its original dir means easier to work out
what the file was exactly).